clean updated

+++ clean.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
-if [ "$(ls -A ./bin)" ]; then
-  rm -r ./bin/*
+if [ -d "./bin" ]; then
+  if [ "$(ls -A ./bin)" ]; then
+    rm -r ./bin/*
+  else
+    echo "The bin directory is already empty."
+  fi
 else
-  echo "The bin directory is already empty."
+  echo "The bin directory does not exist."
 fi
